add conftest for checkpoints tests
